Upgrade em UsuariosController
Agora as senhas são gravadas com hash;
Há tratamentos pra função UsuarioController@show se ela não retornar nenhum resultado;
Função de login criada, porém ainda sem funcionalidade;
Tratamento para quando já existe um login igual no banco;

+++ app/Http/Controllers/Api/UsuariosController.php
@@ -5,7 +5,9 @@ namespace App\Http\Controllers\Api;
 use App\API\ApiError;
 use App\Http\Controllers\Controller;
 use App\Usuarios;
+use Facade\FlareClient\Api;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Hash;
 
 class UsuariosController extends Controller
 {
@@ -27,7 +29,8 @@ class UsuariosController extends Controller
     {
         try {
             $data = ['data' => $id];
-            return response()->json($data);
+            return \response()->json($data);
+            //Tratamento de usuario não encontrado colocado dentro de app\Exceptions\Handler.php
         } catch (\Exception $e) {
             if (config('app.debug')) {
                 return response()->json(ApiError::errorMessage($e->getMessage(), 1010));
@@ -36,14 +39,32 @@ class UsuariosController extends Controller
         }
     }
 
+    public function login(Request $request)
+    {
+        return response()->json(['data' => ['msg' => "sucesso"]], 201);
+        // try {
+        //     // $usuarioData = $request->all();
+        //     // \print_r($usuarioData);
+        //     return response()->json(['data' => ['msg' => "sucesso"]], 201);
+        // } catch (\Exception $e) {
+        //     if (config('app.debug')) {
+        //         return response()->json(ApiError::errorMessage($e->getMessage(), 1010));
+        //     }
+        //     return response()->json(ApiError::errorMessage('Houve um erro ao realizar a operação  de ' . __FUNCTION__, 1010));
+        // }
+    }
+
     public function criar(Request $request)
     {
         try {
-            // dd($request->all());
             $usuarioData = $request->all();
+            $usuarioData['senha'] = Hash::make($usuarioData['senha']);
             $this->usuario->create($usuarioData);
             return response()->json(['data' => ['msg' => "Usuario criado com sucesso"]], 201);
         } catch (\Exception $e) {
+            if ($e->getCode() == 23000) {
+                return response()->json(ApiError::errorMessage("Já existe um usuario com esse login", 422), 422);
+            }
             if (config('app.debug')) {
                 return response()->json(ApiError::errorMessage($e->getMessage(), 1010));
             }
@@ -54,15 +75,19 @@ class UsuariosController extends Controller
     public function alterar(Request $request)
     {
         try {
-            // dd($request->all());
             $usuarioData = $request->all();
             $usuario_encontrado = $this->usuario->find($usuarioData['id']);
-            // print_r($usuario_encontrado);
+            if (isset($usuarioData['senha'])) {
+                $usuarioData['senha'] = Hash::make($usuarioData['senha']);
+            }
             $usuario_encontrado->update($usuarioData);
             return response()->json(['data' => ['msg' => "Usuario alterado com sucesso"]], 201);
         } catch (\Exception $e) {
+            if ($e->getCode() == 23000) {
+                return response()->json(ApiError::errorMessage("Já existe um usuario com esse login", 422), 422);
+            }
             if (config('app.debug')) {
-                return response()->json(ApiError::errorMessage($e->getMessage(), 1010));
+                return response()->json(ApiError::errorMessage($e->getMessage(), 1010), 1010);
             }
             return response()->json(ApiError::errorMessage('Houve um erro ao realizar a operação  de ' . __FUNCTION__, 1010));
         }
@@ -70,13 +95,19 @@ class UsuariosController extends Controller
 
     public function deletar(Request $request)
     {
+        /*
+        Ao deletar um usuario, ele deletará todo o conteudo relacionado a esse usuario
+        como por exemplo os clientes, e os animais referentes a esses clientes
+        */
         try {
             // dd($request->all());
             $usuarioData = $request->all();
             $usuario_encontrado = $this->usuario->find($usuarioData['id']);
-            // print_r($usuario_encontrado);
-            $usuario_encontrado->delete($usuarioData);
-            return response()->json(['data' => ['msg' => "Usuario " . $request['id'] . " deletado com sucesso"]], 201);
+            if (isset($usuario_encontrado)) {
+                $usuario_encontrado->delete($usuarioData);
+                return response()->json(['data' => ['msg' => "Usuario " . $request['id'] . " deletado com sucesso"]], 201);
+            }
+            return response()->json(ApiError::errorMessage("Usuario de id $usuarioData[id] nao encontrado", 404));
         } catch (\Exception $e) {
             if (config('app.debug')) {
                 return response()->json(ApiError::errorMessage($e->getMessage(), 1010));
